requirements.txt
openai
tweepy
python-decouple
pytz
